Add files via upload
This is a program to calculate last digit of the sum of fibonacci number in a range of m and n. This means Sum= Fm+F(m+1)+F(m+2)+......+Fn. Here m<=n and (0<=m<=n<=10^18).
We calculate it efficiently using Pisano's period and we subtract the last digits of sum(Fm-1) with last digit of sum(Fn).